Move component to folfer
